requirements.txt
alabaster==0.7.12
altgraph==0.17
anyio==3.6.1
asks==3.0.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
certifi==2020.12.5
cffi==1.14.5
cfgrib==0.9.9.0
cftime==1.4.1
chardet==4.0.0
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Cython==0.29.28
dcor==0.5.7
dill==0.3.5.1
docutils==0.17.1
eccodes==1.3.2
eccodes-python==0.9.9
ecmwf-api-client==1.5.0
EMD-signal==0.2.15
et-xmlfile==1.1.0
ewtpy==0.2
findlibs==0.0.2
fonttools==4.37.1
future==0.18.2
gensim==4.2.0
h11==0.13.0
idna==2.10
imagesize==1.2.0
importlib-metadata==4.12.0
iniconfig==1.1.1
Jinja2==3.0.2
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.38.0
MarkupSafe==2.0.1
matlabengineforpython===R2020b
matplotlib==3.4.1
metrics==0.3.3
multiprocess==0.70.13
netCDF4==1.5.6
netdata-pandas==0.0.41
numba==0.55.1
numpy==1.20.0
numpydoc==1.1.0
openpyxl==3.0.7
outcome==1.2.0
packaging==21.3
pandas==1.2.4
pathlib2==2.3.7.post1
pathos==0.2.9
pathspec==0.5.5
patsy==0.5.2
pefile==2021.5.13
Pillow==8.2.0
pluggy==1.0.0
pox==0.3.1
ppft==1.7.6.5
prettytable==2.1.0
py==1.10.0
pycparser==2.20
Pygments==2.2.0
pyinstaller-hooks-contrib==2021.1
pyparsing==2.4.7
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
PySide2==5.15.2
pytest==6.2.5
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pywin32==301
pywin32-ctypes==0.2.0
PyYAML==5.4.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.25.1
scikit-learn==0.24.2
scikit-opt==0.6.5
scipy==1.6.3
seaborn==0.11.1
shiboken2==5.15.2
six==1.15.0
sklearn==0.0
smart-open==6.1.0
sniffio==1.3.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
sparse==0.13.0
Sphinx==4.2.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
statsmodels==0.13.2
superflexpy==1.3.1
tabulate==0.8.10
threadpoolctl==2.1.0
toml==0.10.2
tomli==2.0.1
trio==0.21.0
urllib3==1.26.2
vmdpy==0.2
wcwidth==0.2.5
wxPython==4.1.1
xlrd==2.0.1
xlrw==0.0.2
xlutils==2.0.0
xlwings==0.24.9
xlwt==1.3.0
zipp==3.8.1
